+++ src/components/PasswordResetForm/queries.tsx
@@ -2,9 +2,10 @@ import gql from "graphql-tag";
 import { TypedMutation } from "../../core/mutations";
 import { ResetPassword, ResetPasswordVariables } from "./types/ResetPassword";
 
+// fixme: this will be fixed in issue https://github.com/mirumee/saleor-storefront/issues/500
 const passwordResetMutation = gql`
   mutation ResetPassword($email: String!) {
-    customerPasswordReset(input: { email: $email }) {
+    requestPasswordReset(email: $email, redirectUrl: "") {
       errors {
         field
         message